requirements.txt
flask
requests
pandas
python-dotenv
ibm-cos-sdk
botocore
